fixed small bug in logout without token

+++ services/auth/src/controllers/auth.controller.ts
@@ -249,6 +249,10 @@ export const refresh = async (req: Request, res: Response, next: NextFunction) =
 // logout => delete refreshToken from db and from httpOnly cookie
 export const logout = async (req: Request, res: Response) => {
     const { refreshToken } = req.cookies;
+    if (!refreshToken) { // protect the token so deletemany dont wipe out all the db :)
+        res.clearCookie('refreshToken', { path: '/api/auth/refresh' });
+        return res.sendStatus(204);
+    }
     await prisma.refreshToken.deleteMany({ where: { token: refreshToken } });
     res.clearCookie('refreshToken', { path: '/api/auth/refresh' });
     res.sendStatus(204);
